requirements.txt
fastapi
uvicorn
transformers
diffusers
torch
accelerate
huggingface-hub
pillow
python-multipart
pydantic